refactor: splitted files and renamed components to match structure with w3m repo
