+++ src/index.tsx
@@ -1,6 +1,6 @@
 import '@walletconnect/react-native-compat';
 import './config/animations';
 
-export { WalletConnectModal } from './components/WalletConnectModal';
+export { WalletConnectModal } from './modal/wcm-modal';
 export { useWalletConnectModal } from './hooks/useWalletConnectModal';
 export { IProvider, IProviderMetadata } from './types/coreTypes';
