+++ src/partials/wcm-all-wallets-list/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  listContentContainer: {
+    paddingBottom: 12,
+  },
+  loaderContainer: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+  },
+});
