+++ src/partials/wcm-all-wallets-search/styles.ts
@@ -1,0 +1,20 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  listContentContainer: {
+    paddingBottom: 12,
+  },
+  emptyContainer: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  emptyText: {
+    fontSize: 16,
+    fontWeight: '600',
+  },
+  loaderContainer: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+  },
+});
