+++ src/views/wcm-all-wallets-view/styles.ts
@@ -1,0 +1,5 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  searchbar: { marginLeft: 16 },
+});
