+++ src/views/wcm-connect-view/styles.ts
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  explorerContainer: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    alignItems: 'center',
+    marginBottom: 16,
+  },
+  portraitItem: {
+    width: '25%',
+  },
+  landscapeItem: {
+    width: '12.5%',
+  },
+  loaderContainer: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    marginBottom: 16,
+  },
+});
