+++ src/views/wcm-connecting-view/index.tsx
@@ -1,26 +1,27 @@
 import { useCallback, useEffect } from 'react';
-import { Linking, Platform, StyleSheet, View } from 'react-native';
+import { Linking, Platform, View } from 'react-native';
 import { useSnapshot } from 'valtio';
 
-import ModalHeader from '../components/ModalHeader';
-import CopyIcon from '../assets/CopyLarge';
-import useTheme from '../hooks/useTheme';
-import { ToastCtrl } from '../controllers/ToastCtrl';
-import { WcConnectionCtrl } from '../controllers/WcConnectionCtrl';
-import Text from '../components/Text';
-import type { RouterProps } from '../types/routerTypes';
-import { RouterCtrl } from '../controllers/RouterCtrl';
-import Touchable from '../components/Touchable';
-import { UiUtil } from '../utils/UiUtil';
-import RetryIcon from '../assets/Retry';
-import WalletImage from '../components/WalletImage';
-import WalletLoadingThumbnail from '../components/WalletLoadingThumbnail';
-import Chevron from '../assets/Chevron';
-import { CoreHelperUtil } from '../utils/CoreHelperUtil';
-import { StorageUtil } from '../utils/StorageUtil';
-import { AssetUtil } from '../utils/AssetUtil';
-import { ConfigCtrl } from '../controllers/ConfigCtrl';
-import { ApiCtrl } from '../controllers/ApiCtrl';
+import ModalHeader from '../../partials/wcm-modal-header';
+import CopyIcon from '../../assets/CopyLarge';
+import useTheme from '../../hooks/useTheme';
+import { ToastCtrl } from '../../controllers/ToastCtrl';
+import { WcConnectionCtrl } from '../../controllers/WcConnectionCtrl';
+import Text from '../../components/Text';
+import type { RouterProps } from '../../types/routerTypes';
+import { RouterCtrl } from '../../controllers/RouterCtrl';
+import Touchable from '../../components/Touchable';
+import { UiUtil } from '../../utils/UiUtil';
+import RetryIcon from '../../assets/Retry';
+import WalletImage from '../../components/WalletImage';
+import WalletLoadingThumbnail from '../../components/WalletLoadingThumbnail';
+import Chevron from '../../assets/Chevron';
+import { CoreHelperUtil } from '../../utils/CoreHelperUtil';
+import { StorageUtil } from '../../utils/StorageUtil';
+import { AssetUtil } from '../../utils/AssetUtil';
+import { ConfigCtrl } from '../../controllers/ConfigCtrl';
+import { ApiCtrl } from '../../controllers/ApiCtrl';
+import styles from './styles';
 
 function ConnectingView({ onCopyClipboard }: RouterProps) {
   const Theme = useTheme();
@@ -156,74 +157,5 @@ function ConnectingView({ onCopyClipboard }: RouterProps) {
     </>
   );
 }
-
-const styles = StyleSheet.create({
-  walletContainer: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingVertical: 40,
-  },
-  image: {
-    height: 96,
-    width: 96,
-    borderRadius: 28,
-  },
-  continueText: {
-    marginTop: 20,
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  retryButton: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingVertical: 6,
-    paddingHorizontal: 12,
-    borderRadius: 16,
-  },
-  retryIcon: {
-    marginLeft: 4,
-  },
-  text: {
-    color: 'white',
-    fontWeight: '500',
-    fontSize: 14,
-  },
-  alternateText: {
-    marginTop: 16,
-  },
-  getText: {
-    fontSize: 16,
-    fontWeight: '600',
-    marginLeft: 8,
-  },
-  storeText: {
-    fontSize: 14,
-    fontWeight: '600',
-  },
-  footer: {
-    width: '100%',
-  },
-  row: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  upperFooter: {
-    alignItems: 'center',
-    paddingVertical: 16,
-    borderTopWidth: StyleSheet.hairlineWidth,
-    borderBottomWidth: StyleSheet.hairlineWidth,
-  },
-  lowerFooter: {
-    paddingHorizontal: 20,
-    paddingTop: 14,
-    paddingBottom: 24,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  storeIcon: {
-    marginLeft: 4,
-  },
-});
 
 export default ConnectingView;
